Fix flaky test fails due to timeouts on macos. (#458)
<!--- This file has been generated from an external template. Please do
not modify it directly. -->
<!--- Changes should be contributed to
https://github.com/munich-quantum-toolkit/templates. -->

## Description

For some reason, lately, the Z3 solver is slower on macOS 13. This PR
disables running test cases with harder instances for the Z3 solver on
macOS.


## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are focused and
relevant to this change.
- [x] I have added appropriate tests that cover the new/changed
functionality.
- [ ] I have updated the documentation to reflect these changes.
- [ ] I have added entries to the changelog for any noteworthy
additions, changes, fixes, or removals.
- [ ] I have added migration instructions to the upgrade guide (if
needed).
- [x] The changes follow the project's style guidelines and introduce no
new warnings.
- [ ] The changes are fully tested and pass the CI checks.
- [x] I have reviewed my own code changes.