Add functionality to add noise to a Circuit. (#453)
<!--- This file has been generated from an external template. Please do
not modify it directly. -->
<!--- Changes should be contributed to
https://github.com/munich-quantum-toolkit/templates. -->

## Description

Currently, QECC constructs noisy circuits ad-hoc wherever it is needed.
I have rewritten this functionality 10 times in various scenarios
already. This PR adds some functionality to add noise to stim circuits.

This PR does not adapt the simulation classes to use these noise models.
PR #377 includes major changes to the structure of the simulation class.
Refactoring the generation of noisy circuits will be handled by that PR.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are focused and
relevant to this change.
- [x] I have added appropriate tests that cover the new/changed
functionality.
- [x] I have updated the documentation to reflect these changes.
- [x] I have added entries to the changelog for any noteworthy
additions, changes, fixes, or removals.
- [ ] I have added migration instructions to the upgrade guide (if
needed).
- [ ] The changes follow the project's style guidelines and introduce no
new warnings.
- [x] The changes are fully tested and pass the CI checks.
- [x] I have reviewed my own code changes.