Add composition of multiple STIM circuits with defined alignment (#518)
## Description

This PR adds the `compose_compact_stim_circuits` function to compose
multiple circuits and align them either based on the circuit start or
end using `compact_stim_circuit`.
It also adds a parameter to `compact_stim_circuit` to allow
as-soon-as-possible and as-late-as-possible scheduling.

Thought this might be useful (at least for me it is).

## Checklist:

- [x] The pull request only contains commits that are focused and
relevant to this change.
- [x] I have added appropriate tests that cover the new/changed
functionality.
- [ ] I have updated the documentation to reflect these changes.
- [ ] I have added entries to the changelog for any noteworthy
additions, changes, fixes, or removals.
- [x] The changes follow the project's style guidelines and introduce no
new warnings.
- [x] The changes are fully tested and pass the CI checks.
- [x] I have reviewed my own code changes.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>